move into subdir to match pattern
